Add support for graceful poweroff/reboot
Now that erlinit provides seconds between the time that a reboot or
poweroff is initiated and the time that it starts brutally killing OS
processes, the Erlang VM can be powered down gracefully. This adds a
call to :init.stop to make this happen. Since control isn't expected to
be returned after calling reboot, halt, and poweroff, it sleeps and then
System.halt's as a last ditch effort to exit the VM.